+++ lib/gecko/record/contact.rb
@@ -3,7 +3,7 @@ require 'gecko/record/base'
 module Gecko
   module Record
     class Contact < Base
-      belongs_to :company
+      belongs_to :company, writeable_on: :create
 
       attribute :email,        String
       attribute :first_name,   String
@@ -14,9 +14,11 @@ module Gecko
       attribute :phone_number, String
       attribute :fax,          String
       attribute :position,     String
-      attribute :phone,        String
 
       attribute :status,       String,  readonly: true
+
+      ## DEPRECATED
+      # attribute :phone,        String
     end
 
     class ContactAdapter < BaseAdapter