+++ lib/gecko/record/fulfillment.rb
@@ -4,9 +4,11 @@ require 'gecko/record/fulfillment_line_item'
 module Gecko
   module Record
     class Fulfillment < Base
-      belongs_to :order
+      belongs_to :order, writeable_on: :create
       belongs_to :shipping_address, class_name: 'Address'
       belongs_to :billing_address,  class_name: 'Address'
+      belongs_to :stock_location,  class_name: 'Location'
+
       has_many :fulfillment_line_items
 
       attribute :status,           String
@@ -16,11 +18,11 @@ module Gecko
       attribute :notes,            String
       attribute :tracking_url,     String
       attribute :tracking_company, String
-      attribute :destination_url,  String,  readonly: true
 
       attribute :packed_at,        Date
       attribute :shipped_at,       DateTime
       attribute :received_at,      DateTime
+      attribute :service,          String
 
       attribute :receipt,          Hash
     end