+++ lib/gecko/record/order.rb
@@ -8,14 +8,15 @@ module Gecko
       has_many :order_line_items
 
       belongs_to :company
+      belongs_to :contact
       belongs_to :shipping_address,     class_name: 'Address'
       belongs_to :billing_address,      class_name: 'Address'
-      belongs_to :contact,              class_name: 'Contact'
       belongs_to :user,                 readonly: true
       belongs_to :assignee,             class_name: 'User'
       belongs_to :stock_location,       class_name: 'Location'
       belongs_to :currency
       # belongs_to :default_price_list,   class_name: 'PriceList'
+      attribute :default_price_list_id, String
 
       attribute :order_number,          String
       attribute :phone_number,          String
@@ -26,8 +27,12 @@ module Gecko
       attribute :packed_status,         String,     readonly: true
       attribute :fulfillment_status,    String,     readonly: true
       attribute :invoice_status,        String,     readonly: true
-      attribute :payment_status,        String
-      attribute :tax_type,              String
+      attribute :payment_status,        String,     readonly: true
+      attribute :return_status,         String,     readonly: true
+      attribute :returning_status,      String,     readonly: true
+      attribute :shippability_status,   String,     readonly: true
+      attribute :backordering_status,   String,     readonly: true
+      attribute :tax_treatment,         String
       attribute :issued_at,             Date
       attribute :ship_at,               Date
       attribute :tags,                  Array[String]
@@ -41,7 +46,9 @@ module Gecko
 
       ## DEPRECATED
       attribute :tracking_number,       String,     readonly: true
+
       # attribute :source,                String
+
       # attribute :invoice_numbers,       Hash[Integer => String], readonly: true
     end