+++ lib/gecko/record/purchase_order.rb
@@ -12,28 +12,29 @@ module Gecko
       belongs_to :billing_address,    class_name: "Location"
 
       belongs_to :currency
-      belongs_to :default_price_list, class_name: "PriceList"
+      # belongs_to :default_price_list, class_name: "PriceList"
+      attribute :default_price_list_id, String
 
       attribute :order_number,          String
       attribute :reference_number,      String
       attribute :email,                 String
       attribute :due_at,                Date
+      attribute :received_at,           Date
 
       attribute :status,                String
-      attribute :procurement_status,    String
+      attribute :procurement_status,    String,     readonly: true
       attribute :notes,                 String
       attribute :tax_treatment,         String
 
-      attribute :destination_url,       String
-      attribute :document_url,          String
+      attribute :document_url,          String,     readonly: true
 
-      attribute :total,                 BigDecimal
-      attribute :cached_quantity,       BigDecimal
-      attribute :cached_total,          BigDecimal
+      attribute :total,                 BigDecimal, readonly: true
+      attribute :cached_quantity,       BigDecimal, readonly: true
+
+      attribute :tags,                  Array[String]
 
       # DEPRECATED
-      # attribute :tax_type, String
-      # attribute :default_price_type_id
+      # attribute :cached_total,          BigDecimal, readonly: true
     end
 
     class PurchaseOrderAdapter < BaseAdapter