+++ lib/gecko/record/purchase_order_line_item.rb
@@ -3,8 +3,8 @@ require 'gecko/record/base'
 module Gecko
   module Record
     class PurchaseOrderLineItem < Base
-      belongs_to :purchase_order
-      belongs_to :variant
+      belongs_to :purchase_order, writeable_on: :create
+      belongs_to :variant, writeable_on: :create
       belongs_to :procurement
       belongs_to :tax_type
 
@@ -14,10 +14,10 @@ module Gecko
       attribute :price,             BigDecimal
       attribute :label,             String
       attribute :freeform,          Boolean
-      attribute :base_price,        BigDecimal
-      attribute :extra_cost_value,  BigDecimal
-      attribute :image_url,         String
 
+      attribute :base_price,        BigDecimal, readonly: true
+      attribute :extra_cost_value,  BigDecimal, readonly: true
+      attribute :image_url,         String,     readonly: true
       # DEPRECATED
       # attribute :tax_rate, String
     end