+++ test/support/shared_adapter_examples.rb
@@ -66,7 +66,7 @@ module SharedAdapterExamples
   end
 
   def test_fetch_miss
-    request_stub = stub_request(:get, /#{plural_name}\/\d+/)
+    stub_request(:get, /#{plural_name}\/\d+/)
       .to_return({
         status:  404,
         headers: {"Content-Type" => "application/json"},
@@ -145,12 +145,29 @@ module SharedAdapterExamples
     assert(record.errors[:title].any?)
   end
 
+  def test_saving_record_with_idempotency_key
+    record = adapter.build
+    mock_token = mock
+    mock_response = mock(status: 200, parsed: {plural_name.singularize => {id: 123}})
+    mock_token.expects(:request)
+              .with(:post, plural_name, {
+                body: record.as_json.to_json,
+                raise_errors: false,
+                headers: {
+                  'Content-Type' => 'application/json',
+                  'Idempotency-Key' => 'abcdefghijkl'
+                }
+              }).returns(mock_response)
+    adapter.client.access_token = mock_token
+    adapter.save(record, idempotency_key: 'abcdefghijkl')
+  end
+
 private
   def mock_api_request(record, request, response)
     mock_token = mock
     mock_response = mock(status: response[0], parsed: response[1])
     mock_token.expects(:request)
-              .with(request[0], request[1], body: record.as_json, raise_errors: false)
+              .with(request[0], request[1], body: record.as_json.to_json, raise_errors: false, headers: {'Content-Type' => 'application/json'})
               .returns(mock_response)
     adapter.client.access_token = mock_token
   end